Add pausing functionality to the game

See merge request Quellus/libgdx-tower-defense!7

+++ core/src/com/quellus/libgdxgame/Input.java
@@ -15,7 +15,7 @@ public class Input {
     this.menu = menu;
   }
 
-  public void handleInput(int x, int y) {
+  public void handleInput(float x, float y) {
     if (!isSelected) {
       isSelected = menu.select(x, y);
       isSelected = true;
